+++ Controller/JsloaderController.php
@@ -14,7 +14,9 @@ namespace Symfony\Cmf\Bundle\CreateBundle\Controller;
 
 use FOS\RestBundle\View\ViewHandlerInterface,
     FOS\RestBundle\View\View;
+use Symfony\Cmf\Bundle\CreateBundle\Security\AccessCheckerInterface;
 use Symfony\Cmf\Bundle\MediaBundle\File\BrowserFileHelper;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\SecurityContextInterface;
 
@@ -29,19 +31,14 @@ use Symfony\Component\Security\Core\SecurityContextInterface;
 class JsloaderController
 {
     /**
-     * @var SecurityContextInterface
-     */
-    protected $securityContext;
-
-    /**
      * @var ViewHandlerInterface
      */
     protected $viewHandler;
 
     /**
-     * @var string the role name for the security check
+     * @var AccessCheckerInterface
      */
-    protected $requiredRole;
+    protected $accessChecker;
 
     /**
      * @var string
@@ -78,6 +75,7 @@ class JsloaderController
      * Create the Controller
      *
      * @param ViewHandlerInterface     $viewHandler
+     * @param AccessCheckerInterface   $accessChecker
      * @param string                   $stanbolUrl         the url to use for
      *      the semantic enhancer stanbol.
      * @param Boolean                  $imageUploadEnabled used to determine
@@ -98,23 +96,21 @@ class JsloaderController
      */
     public function __construct(
         ViewHandlerInterface $viewHandler,
-        $stanbolUrl,
+        AccessCheckerInterface $accessChecker,
+        $stanbolUrl = false,
         $imageUploadEnabled = false,
         $fixedToolbar = true,
         $plainTextTypes = array(),
-        $requiredRole = "IS_AUTHENTICATED_ANONYMOUSLY",
-        SecurityContextInterface $securityContext = null,
         $editorBasePath = null,
         BrowserFileHelper $browserFileHelper = null
     ) {
         $this->viewHandler        = $viewHandler;
+        $this->accessChecker      = $accessChecker;
         $this->stanbolUrl         = $stanbolUrl;
         $this->imageUploadEnabled = $imageUploadEnabled;
         $this->fixedToolbar       = $fixedToolbar;
         $this->plainTextTypes     = $plainTextTypes;
         $this->editorBasePath     = $editorBasePath;
-        $this->requiredRole       = $requiredRole;
-        $this->securityContext    = $securityContext;
         $this->browserFileHelper  = $browserFileHelper;
     }
 
@@ -130,11 +126,12 @@ class JsloaderController
      *   CmfCreateBundle::includejsfiles-%editor%.html.twig
      * and pass the appropriate editor name.
      *
-     * @param string $editor the name of the editor to load.
+     * @param Request $request The request object for the AccessChecker.
+     * @param string  $editor  the name of the editor to load.
      */
-    public function includeJSFilesAction($editor = 'ckeditor')
+    public function includeJSFilesAction(Request $request, $editor = 'ckeditor')
     {
-        if (!$this->performSecurityChecks()) {
+        if (!$this->accessChecker->check($request)) {
             return new Response('');
         }
 
@@ -161,29 +158,5 @@ class JsloaderController
         );
 
         return $this->viewHandler->handle($view);
-    }
-
-    /**
-     * Actions may be performed if the requiredRole is set to false (completely
-     * disable security check) or if there is a securityContext and it grants
-     * the required role.
-     *
-     * @return boolean Whether the current user is allowed to edit.
-     */
-    protected function performSecurityChecks()
-    {
-        if (false === $this->requiredRole) {
-            // security check is disabled
-            return true;
-        }
-
-        if ($this->securityContext
-            && $this->securityContext->getToken()
-            && $this->securityContext->isGranted($this->requiredRole)
-        ) {
-            return true;
-        }
-
-        return false;
     }
 }